\name{pel-package}
\alias{pel-package}
\alias{pel}
\docType{package}
\title{
\packageTitle{pel}
}
\description{
\packageDescription{pel}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{pel}
\packageIndices{pel}
An overview of how to use the package, including the most important
functions
}
\author{
\packageAuthor{pel}

Maintainer: \packageMaintainer{pel}
}
\references{
Literature or other references for background information
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
\keyword{ package }
\seealso{TODO
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{a <- "TODO"
%%~~ simple examples of the most important functions ~~
}
